Base project clean up; Rewrite building scripts; Delete stb files i don't need; Delete nucklear(why i didn't i do it earlier?); More Math functions are done (replace CGLM with my Math in Engine, will use it only in SimpleTest); Clean code for editor

+++ Engine/src/Math/V4.h
@@ -29,7 +29,7 @@ v4_copy(v4 a, v4 b)
 force_inline f32
 v4_length(v4 v)
 {
-    return sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
+    return sqrtf(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
 }
 
 force_inline f32
